#264 renamed opening search dialogs
