+++ src/components/Chess.js
@@ -25,9 +25,9 @@ import LoadPgnDialog from "./Dialog/LoadPgnDialog";
 import RematchAcceptDialog from "./Dialog/RematchAcceptDialog";
 import RematchOfferDialog from "./Dialog/RematchOfferDialog";
 import PlayLikeGrandmasterDialog from "./Dialog/PlayLikeGrandmasterDialog";
-import OpeningSearchEcoDialog from "./Dialog/OpeningSearchEcoDialog";
-import OpeningSearchNameDialog from "./Dialog/OpeningSearchNameDialog";
-import OpeningSearchMovetextDialog from "./Dialog/OpeningSearchMovetextDialog";
+import OpeningSearchEcoDialog from "./Dialog/ChessOpeningSearchEcoDialog";
+import OpeningSearchNameDialog from "./Dialog/ChessOpeningSearchNameDialog";
+import OpeningSearchMovetextDialog from "./Dialog/ChessOpeningSearchMovetextDialog";
 import Board from "./Board.js";
 import Buttons from "./Buttons.js";
 import History from "./History";
